requirements.txt
gast
imageio>=2.12.0
imgaug
lmdb
lxml>=4.6.4
matplotlib>=3.5.0
numpy>=1.21.4
opencv-contrib-python>=4.4.0.46
opencv-python>=4.5.3.56
openpyxl
paddlepaddle-gpu~=2.2.1
Pillow>=8.4.0
pyclipper
pyparsing
pytest>=6.2.5
python-dotenv>=0.19.2
python-Levenshtein
PyYAML>=6.0
requests>=2.26.0
scikit-image>=0.17.2
scipy>=1.7.3
setuptools-scm
Shapely>=1.8.0
tqdm
urllib3>=1.26.7
virtualenv>=20.10.0
Werkzeug
visualdl
protobuf~=3.20
GPUtil
psutil
pynvml
distro
wheel
onnx
onnxruntime-gpu~=1.9.0
